Intitial stages of building Construction WORKING
Keenos now, when idle, check to see if any buildings are UnderConstruction.
If they find one, they check if the player has the necessary resources.
If they do, they pick up one unit of said resource and deliver it.
IMPORTANT:
Currently, multiple Keeno will be able to bring resources to the building if the player can afford it. This is fine until 10 Keeno bring 1 resource each to the same building when it's only missing 1. The code does not currently account for that.